# Documentazione/OpenAPI/openapi.yaml
openapi: 3.0.3
info:
  title: BugBoard26 API
  version: 1.0.0
  description: API surface for BugBoard26.JWT session revocation remains enabled as
    an intentional security exception.
paths:
...
  /api/issues/{issueId}/events:
    get:
      operationId: issues_events_list
      description: Issue activity events.
      parameters:
      - in: path
        name: issueId
        schema:
          type: integer
        description: A unique integer value identifying this issue.
        required: true
      tags:
      - Issues
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IssueEvent'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueEvent'
          description: ''
    post:
      operationId: issues_events_create
      description: Issue activity events.
      parameters:
      - in: path
        name: issueId
        schema:
          type: integer
        description: A unique integer value identifying this issue.
        required: true
      tags:
      - Issues
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueUpdateJsonRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IssueUpdateMultipartRequestRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IssueEvent'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueEvent'
          description: ''
  /api/issues/{issueId}/events/stream:
    get:
      operationId: issues_events_stream_retrieve
      description: Server-Sent Events activity stream for the issue.
      parameters:
      - in: path
        name: issueId
        schema:
          type: integer
        description: A unique integer value identifying this issue.
        required: true
      tags:
      - Issues
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            text/event-stream:
              schema:
                type: string
          description: ''
  /api/issues/{issueId}/subscriptions/me:
    get:
      operationId: issues_subscriptions_me_retrieve
      description: Admin subscription state for the authenticated user on the issue.
      parameters:
      - in: path
        name: issueId
        schema:
          type: integer
        description: A unique integer value identifying this issue.
        required: true
      tags:
      - Issues
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueSubscriptionState'
          description: ''
        '204':
          description: Subscription updated
    put:
      operationId: issues_subscriptions_me_update
      description: Admin subscription state for the authenticated user on the issue.
      parameters:
      - in: path
        name: issueId
        schema:
          type: integer
        description: A unique integer value identifying this issue.
        required: true
      tags:
      - Issues
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueSubscriptionState'
          description: ''
        '204':
          description: Subscription updated
    delete:
      operationId: issues_subscriptions_me_destroy
      description: Admin subscription state for the authenticated user on the issue.
      parameters:
      - in: path
        name: issueId
        schema:
          type: integer
        description: A unique integer value identifying this issue.
        required: true
      tags:
      - Issues
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueSubscriptionState'
          description: ''
        '204':
          description: Subscription updated
  /api/issues/{issueId}/suggestions:
    get:
      operationId: issues_suggestions_list
      parameters:
      - in: path
        name: issueId
        schema:
          type: integer
        description: A unique integer value identifying this issue.
        required: true
      tags:
      - Issues
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IssueSuggestion'
          description: ''
...
components:
  schemas:
...
    Attachment:
      type: object
      properties:
        attachmentId:
          type: integer
          readOnly: true
        updateId:
          type: integer
          readOnly: true
        originalName:
          type: string
          readOnly: true
        path:
          type: string
          maxLength: 256
        url:
          type: string
          readOnly: true
        mimeType:
          type: string
        size:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        uploadedAt:
          type: string
          format: date-time
          readOnly: true
      required:
      - attachmentId
      - mimeType
      - originalName
      - path
      - size
      - updateId
      - uploadedAt
      - url
...
    IssueEvent:
      type: object
      properties:
        updateId:
          type: integer
          readOnly: true
        issueId:
          type: integer
          readOnly: true
        actorId:
          type: integer
          readOnly: true
        actorUsername:
          type: string
          readOnly: true
        eventType:
          type: string
        at:
          type: string
          format: date-time
          readOnly: true
        message:
          type: string
          maxLength: 1000
        oldStatus:
          type: string
        newStatus:
          type: string
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          readOnly: true
      required:
      - actorId
      - actorUsername
      - at
      - attachments
      - eventType
      - issueId
      - updateId
...
    IssueSubscriptionState:
      type: object
      properties:
        subscribed:
          type: boolean
      required:
      - subscribed
    IssueSuggestion:
      type: object
      properties:
        projectMembershipId:
          type: integer
          readOnly: true
        projectId:
          type: integer
          readOnly: true
        userId:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
        firstName:
          type: string
          readOnly: true
        lastName:
          type: string
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        role:
          type: string
          readOnly: true
        profileImg:
          type: string
          readOnly: true
        openCount:
          type: integer
      required:
      - email
      - firstName
      - lastName
      - openCount
      - profileImg
      - projectId
      - projectMembershipId
      - role
      - userId
      - username
    IssueUpdateJsonRequestRequest:
      type: object
      properties:
        message:
          type: string
          minLength: 1
      required:
      - message
    IssueUpdateMultipartRequestRequest:
      type: object
      properties:
        message:
          type: string
          minLength: 1
        file:
          type: array
          items:
            type: string
            format: binary
      required:
      - message
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRFToken